Removed all unnecessary interfaces from strategies
Removed all interfaces from the strategy classes that only have one
implementation.
