+++ labs/abiquo/src/main/java/org/jclouds/abiquo/strategy/cloud/ListVirtualDatacenters.java
@@ -19,28 +19,98 @@
 
 package org.jclouds.abiquo.strategy.cloud;
 
-import java.util.List;
+import static com.google.common.base.Preconditions.checkNotNull;
+import static com.google.common.collect.Iterables.filter;
+import static org.jclouds.abiquo.domain.DomainWrapper.wrap;
+import static org.jclouds.concurrent.FutureIterables.transformParallel;
 
+import java.util.List;
+import java.util.concurrent.ExecutorService;
+import java.util.concurrent.Future;
+
+import javax.annotation.Resource;
+import javax.inject.Named;
+
+import org.jclouds.Constants;
+import org.jclouds.abiquo.AbiquoApi;
+import org.jclouds.abiquo.AbiquoAsyncApi;
+import org.jclouds.abiquo.domain.DomainWrapper;
 import org.jclouds.abiquo.domain.cloud.VirtualDatacenter;
 import org.jclouds.abiquo.domain.cloud.options.VirtualDatacenterOptions;
 import org.jclouds.abiquo.strategy.ListRootEntities;
-import org.jclouds.abiquo.strategy.cloud.internal.ListVirtualDatacentersImpl;
+import org.jclouds.logging.Logger;
+import org.jclouds.rest.RestContext;
 
+import com.abiquo.server.core.cloud.VirtualDatacenterDto;
+import com.abiquo.server.core.cloud.VirtualDatacentersDto;
+import com.google.common.base.Function;
 import com.google.common.base.Predicate;
-import com.google.inject.ImplementedBy;
+import com.google.common.collect.Lists;
+import com.google.inject.Inject;
+import com.google.inject.Singleton;
 
 /**
  * List virtual datacenters.
  * 
  * @author Ignasi Barrera
  */
-@ImplementedBy(ListVirtualDatacentersImpl.class)
-public interface ListVirtualDatacenters extends ListRootEntities<VirtualDatacenter> {
-   Iterable<VirtualDatacenter> execute(VirtualDatacenterOptions virtualDatacenterOptions);
+@Singleton
+public class ListVirtualDatacenters implements ListRootEntities<VirtualDatacenter> {
+   protected final RestContext<AbiquoApi, AbiquoAsyncApi> context;
 
-   Iterable<VirtualDatacenter> execute(List<Integer> virtualDatacenterIds);
+   protected final ExecutorService userExecutor;
 
-   Iterable<VirtualDatacenter> execute(Predicate<VirtualDatacenter> selector,
-         VirtualDatacenterOptions virtualDatacenterOptions);
+   @Resource
+   protected Logger logger = Logger.NULL;
+
+   @Inject(optional = true)
+   @Named(Constants.PROPERTY_REQUEST_TIMEOUT)
+   protected Long maxTime;
+
+   @Inject
+   ListVirtualDatacenters(final RestContext<AbiquoApi, AbiquoAsyncApi> context,
+         @Named(Constants.PROPERTY_USER_THREADS) final ExecutorService userExecutor) {
+      this.context = checkNotNull(context, "context");
+      this.userExecutor = checkNotNull(userExecutor, "userExecutor");
+   }
+
+   @Override
+   public Iterable<VirtualDatacenter> execute() {
+      VirtualDatacenterOptions virtualDatacenterOptions = VirtualDatacenterOptions.builder().build();
+
+      return execute(virtualDatacenterOptions);
+   }
+
+   @Override
+   public Iterable<VirtualDatacenter> execute(final Predicate<VirtualDatacenter> selector) {
+      return filter(execute(), selector);
+   }
+
+   public Iterable<VirtualDatacenter> execute(final VirtualDatacenterOptions virtualDatacenterOptions) {
+      VirtualDatacentersDto result = context.getApi().getCloudApi().listVirtualDatacenters(virtualDatacenterOptions);
+      return wrap(context, VirtualDatacenter.class, result.getCollection());
+   }
+
+   public Iterable<VirtualDatacenter> execute(final Predicate<VirtualDatacenter> selector,
+         final VirtualDatacenterOptions virtualDatacenterOptions) {
+      return filter(execute(virtualDatacenterOptions), selector);
+   }
+
+   public Iterable<VirtualDatacenter> execute(final List<Integer> virtualDatacenterIds) {
+      // Find virtual datacenters in concurrent requests
+      return listConcurrentVirtualDatacenters(virtualDatacenterIds);
+   }
+
+   private Iterable<VirtualDatacenter> listConcurrentVirtualDatacenters(final List<Integer> ids) {
+      Iterable<VirtualDatacenterDto> vdcs = transformParallel(ids,
+            new Function<Integer, Future<? extends VirtualDatacenterDto>>() {
+               @Override
+               public Future<VirtualDatacenterDto> apply(final Integer input) {
+                  return context.getAsyncApi().getCloudApi().getVirtualDatacenter(input);
+               }
+            }, userExecutor, maxTime, logger, "getting virtual datacenters");
+
+      return DomainWrapper.wrap(context, VirtualDatacenter.class, Lists.newArrayList(vdcs));
+   }
 
 }
