+++ labs/abiquo/src/test/java/org/jclouds/abiquo/strategy/enterprise/ListVirtualMachineTemplatesLiveApiTest.java
@@ -17,49 +17,54 @@
  * under the License.
  */
 
-package org.jclouds.abiquo.strategy.infrastructure.internal;
+package org.jclouds.abiquo.strategy.enterprise;
 
 import static com.google.common.collect.Iterables.size;
 import static org.testng.Assert.assertEquals;
 import static org.testng.Assert.assertNotNull;
 import static org.testng.Assert.assertTrue;
 
-import org.jclouds.abiquo.domain.infrastructure.Machine;
-import org.jclouds.abiquo.predicates.infrastructure.MachinePredicates;
+import org.jclouds.abiquo.domain.cloud.VirtualMachineTemplate;
+import org.jclouds.abiquo.predicates.cloud.VirtualMachineTemplatePredicates;
 import org.jclouds.abiquo.strategy.BaseAbiquoStrategyLiveApiTest;
+import org.jclouds.abiquo.strategy.cloud.ListVirtualAppliances;
+import org.jclouds.abiquo.strategy.enterprise.ListVirtualMachineTemplates;
 import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Test;
 
 /**
- * Live tests for the {@link ListMachinesImpl} strategy.
+ * Live tests for the {@link ListVirtualAppliances} strategy.
  * 
  * @author Ignasi Barrera
  */
-@Test(groups = "api", testName = "ListMachinesImplLiveApiTest")
-public class ListMachinesImplLiveApiTest extends BaseAbiquoStrategyLiveApiTest {
-   private ListMachinesImpl strategy;
+@Test(groups = "api", testName = "ListVirtualMachineTemplatesLiveApiTest")
+public class ListVirtualMachineTemplatesLiveApiTest extends BaseAbiquoStrategyLiveApiTest {
+   private ListVirtualMachineTemplates strategy;
 
    @Override
    @BeforeClass(groups = "api")
    protected void setupStrategy() {
-      this.strategy = env.context.getUtils().getInjector().getInstance(ListMachinesImpl.class);
+      this.strategy = env.context.getUtils().getInjector().getInstance(ListVirtualMachineTemplates.class);
    }
 
    public void testExecute() {
-      Iterable<Machine> machines = strategy.execute();
-      assertNotNull(machines);
-      assertTrue(size(machines) > 0);
+      Iterable<VirtualMachineTemplate> templates = strategy.execute(env.defaultEnterprise);
+      assertNotNull(templates);
+      assertTrue(size(templates) > 0);
    }
 
    public void testExecutePredicateWithoutResults() {
-      Iterable<Machine> machines = strategy.execute(MachinePredicates.name("UNEXISTING"));
-      assertNotNull(machines);
-      assertEquals(size(machines), 0);
+      Iterable<VirtualMachineTemplate> templates = strategy.execute(env.defaultEnterprise,
+            VirtualMachineTemplatePredicates.name("UNEXISTING"));
+      assertNotNull(templates);
+      assertEquals(size(templates), 0);
    }
 
    public void testExecutePredicateWithResults() {
-      Iterable<Machine> machines = strategy.execute(MachinePredicates.name(env.machine.getName()));
-      assertNotNull(machines);
-      assertEquals(size(machines), 1);
+      Iterable<VirtualMachineTemplate> templates = strategy.execute(env.defaultEnterprise,
+            VirtualMachineTemplatePredicates.name(env.template.getName()));
+      assertNotNull(templates);
+      // Repository can have multiple templates with the same name
+      assertTrue(size(templates) > 0);
    }
 }
